l10n: adding the basic l10n function
See the deps/l10n/README.md

Adds a basically l10n function based on ICU's resource bundle
mechanism. Works for both the C/C++ and JS core library. This
is only intended for use with Node's own messages and outputs.
This is intended for initial review. I'm certain improvements can be
made.

To enable this, configure the node with the --with-intl switch, just
like we do with the Intl support. If --with-intl is not switched on,
then the l10n falls back to simple printf statements. When --with-intl
is switched on, the resources are compiled into a static library and
linked statically into node.